requirements.txt
appnope==0.1.0
boto==2.48.0
bz2file==0.98
certifi==2017.7.27.1
chardet==3.0.4
Cython==0.26.1
decorator==4.1.2
idna==2.6
jedi==0.10.2
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psutil==5.3.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
regex==2017.9.23
requests==2.18.4
scipy==0.19.1
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.3
torchvision==0.1.9
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7